feat: adjust css
