+++ src/components/avater_svg/avater_svg.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+
+const AvatarSVG = ({ size }: { size: 'large' | 'small' }) => {
+  const width = size === 'large' ? 201 : 100;
+  const height = width;
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={width}
+      height={height}
+      fill="none"
+      viewBox="0 0 201 200"
+    >
+      <path
+        fill="#CDD1D9"
+        d="M.5 100C.5 44.772 45.272 0 100.5 0s100 44.772 100 100-44.772 100-100 100S.5 155.228.5 100z"
+      ></path>
+      <g clipPath="url(#clip0_13_13411)">
+        <path
+          fill="#7F8A9D"
+          fillRule="evenodd"
+          d="M100.5 68.013c-11.942 0-21.623 9.68-21.623 21.622 0 8.151 4.51 15.249 11.17 18.934a31.953 31.953 0 00-19.976 20.439 2.286 2.286 0 002.177 2.984h56.503a2.284 2.284 0 002.176-2.984 31.956 31.956 0 00-19.975-20.439c6.661-3.685 11.171-10.782 11.171-18.934 0-11.942-9.681-21.622-21.623-21.622z"
+          clipRule="evenodd"
+        ></path>
+      </g>
+      <defs>
+        <clipPath id="clip0_13_13411">
+          <path fill="#fff" d="M0 0H64V64H0z" transform="translate(68.5 68)"></path>
+        </clipPath>
+      </defs>
+    </svg>
+  );
+};
+
+export default AvatarSVG;
