+++ src/components/login_confirm_modal/login_confirm_modal.beta.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
+import { RxCross2 } from 'react-icons/rx';
 import { useTranslation } from 'react-i18next';
-import InformationStatement from './information_statement';
-import TermsOfServiceAndPrivacyPolicy from './term_n_privacy';
+import InformationStatement from '@/components/login_confirm_modal/information_statement';
+import TermsOfServiceAndPrivacyPolicy from '@/components/login_confirm_modal/term_n_privacy';
 
 interface ILoginConfirmProps {
   isModalVisible: boolean;
@@ -23,9 +24,15 @@ const LoginConfirmModal: React.FC<ILoginConfirmProps> = ({
   const { t } = useTranslation('common');
   return isModalVisible ? (
     <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50">
-      <div className="max-h-80vh w-full max-w-xl rounded-xs bg-white p-6 shadow-lg">
-        <h2 className="mb-4 text-lg font-semibold">{modalData.title}</h2>
-        <div className="h-50vh mb-4 overflow-y-auto rounded-xs border border-navyBlue p-4">
+      <div className="absolute max-h-80vh w-full max-w-xl rounded-xs bg-white p-4 pt-0 shadow-lg">
+        <div className="my-4 mb-8 flex items-center justify-between">
+          <h2 className="text-lg font-semibold text-navyBlue">{modalData.title}</h2>
+          <button type="button" onClick={onCancel} className="text-navyBlue">
+            <RxCross2 size={20} />
+          </button>
+        </div>
+        <hr className="absolute left-0 top-60px w-full max-w-xl border-lightGray6" />
+        <div className="h-50vh m-4 overflow-y-auto rounded-xs border border-navyBlue bg-lightGray7 lg:p-4">
           {modalData.content === 'info_collection_statement' && <InformationStatement />}
           {modalData.content === 'term_n_privacy' && <TermsOfServiceAndPrivacyPolicy />}
         </div>
@@ -33,14 +40,14 @@ const LoginConfirmModal: React.FC<ILoginConfirmProps> = ({
           <button
             type="button"
             onClick={onCancel}
-            className="mr-2 rounded-xs border border-navyBlue px-4 py-2 text-navyBlue hover:bg-gray-100"
+            className="mr-2 rounded-xs border border-navyBlue px-4 py-2 text-sm text-navyBlue hover:bg-gray-100 lg:text-lg"
           >
             {t('COMMON.CANCEL')}
           </button>
           <button
             type="button"
             onClick={onAgree}
-            className="rounded-xs bg-navyBlue px-4 py-2 text-white hover:bg-navyBlue2"
+            className="rounded-xs bg-navyBlue px-4 py-2 text-sm text-white hover:bg-navyBlue2 lg:text-lg"
           >
             {modalData.buttonText}
           </button>
