+++ src/components/login_confirm_modal/term_n_privacy.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const TermsOfServiceAndPrivacyPolicy: React.FC = () => {
   return (
-    <div className="p-6 space-y-4 text-gray-800">
+    <div className="p-6 space-y-4 text-navyBlue">
       <p className="font-semibold">Dear User,</p>
       <p>
         Welcome to our accounting software. To ensure you fully understand our
