+++ src/lib/hooks/use_api_worker.ts
@@ -7,8 +7,8 @@ import { STATUS_CODE, STATUS_MESSAGE } from '@/constants/status_code';
 const useAPIWorker = <Data>(
   apiConfig: IAPIConfig,
   options: IAPIInput,
+  triggerImmediately: boolean = false,
   cancel?: boolean,
-  triggerImmediately: boolean = true
 ): IAPIResponse<Data> => {
   const [success, setSuccess] = useState<boolean | undefined>(undefined);
   const [code, setCode] = useState<string | undefined>(undefined);
