+++ tailwind.config.ts
@@ -743,6 +743,7 @@ module.exports = {
         darkBlue3: '#0024620D',
       },
       backgroundImage: {
+        login_bg: "url('/images/login_bg.svg')",
         customGradient: 'linear-gradient(180deg, rgba(0, 24, 64, 0) 10%, rgb(0, 24, 64) 85.5%)',
         contact_bg: 'url("/elements/contact_bg.svg")',
